refactor: 깃허브 피드백 반영 및 예외처리
